Invert parity of chessboard shader.
The original `(p && q) || !(p || q)` line from the second Unity
chessboard sample really means "both or neither". This can be simplified
to `!(p ^ q)` and then the parity of the board can be inverted by
dropping the "not", yielding a simpler `p ^ q` for the condition and
bringing this shader into line with the other two in terms of
"chessboard parity". The end result: easier side-by-side comparison of
the different chessboard shaders in this project without having the
squares flip between light and dark.